refactor waterfall settings panels into components
